+++ CHANGELOG.md
@@ -9,12 +9,13 @@
 * Added struct `OptionalDataOrConstRef` similar to `std::optional` but can hold `std::nullopt`, a type `T` or a const reference `const T&`.
 * Added struct `OptionalRef` similar to `std::optional` but can hold `std::nullopt` or a reference `T&`/`const T&`.
 * Improved control for `Stringify` with Json output.
-* Updated `StringifyOptions` (breaking change, but no field renames).
-* Updated `StringifyWithFieldNames` (breaking change).
+  * Updated `StringifyOptions` (breaking change).
+  * Updated `StringifyWithFieldNames` (breaking change).
 * Modified `MboTypesStringifyOptions` (breaking change) to be more flexible - but users must deal with the change.
 * Added ability to detect bad `MboTypesStringifyOptions` signatures.
 * Added `StringifyFieldOptions` which holds outer and inner `StringifyOptions`.
 * Added `Stringify::AsJsonPretty()` and `StringifyOptions::AsJsonPretty()`.
+* Added `Demangle` to log de-mangled typeid names.
 
 # 0.9.0